requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3.1
python-decouple>=3.8
celery>=5.3.6
redis>=5.0.1
python-telegram-bot>=20.3